chore(bin/domain-sift): update comment
No point in deciding prematurely what approach should be used. I had
talked privately with another dev some time ago and they had suggested a
trie instead. Chances are that I'm not implementing this any time soon,
so the comment is revised to be more general. I still think it's
important to mention this, though.